Create SuperDARN display product PT10M
I've identified the data as "restricted" although this is a mixed
situation.  The "real time" data of the past month is openly
available, while the "data archive" older data needs a user name
and password.  I did not think this distinction required two
separate data product descriptors.

I identified the person named as JHU/APL PI as the general
contact.  There are very many PI's, but I don't see the name
of anyone I recognize as a better general contact (as Ray
Greenwald would have been in the past.)

I guessed at the format of the display images as GIF's.  It's not
clear to me how to know whether they're GIFs or JPEGs or whatever.

Signed-off-by: Jan Merka <jan.merka@nasa.gov>